Move methods to send specific messages from channels to KernelClient
